append for pdf && delete routes

+++ routes/api.js
@@ -34,6 +34,13 @@ router.get('/document/:name',(req,res) => {
   }
   else res.status(404).send("not matching extension");
 });
+
+router.delete('/document/:name',(req,res) => {
+  let name=req.params.name;
+  storageApi.DeleteFile(name, null, null,  (responseMessage) => {
+    res.status(responseMessage.code).send(responseMessage);
+  })
+})
 //download files
 //original(append original): `http://localhost:4200/api/download/originald041b0836d7804a4b40cb137a5bdfddftest1.pptx`
 //modified (normal name): `http://localhost:4200/api/download/d041b0836d7804a4b40cb137a5bdfddftest1.pptx`
@@ -59,35 +66,62 @@ router.get('/download/append/:name', (req,res) => {
   const name = `original${req.params.name}`; //the file we will copy
   const tmpname= `append${req.params.name}`; //the name of the final file
   if(req.params.name.substring(0,8) === 'original') res.status(404).send('use without original in file name');
-  const documentEntryList =  {
-    'DocumentEntries' : [{
-      'Href' : `${req.params.name}`,
-      'ImportFormatMode' : 'KeepSourceFormatting'
-    }
-  ]};
-  if ( wordsRegex.test(path.extname(name))) {
+  if ( wordsRegex.test(path.extname(name)) || pdfRegex.test(path.extname(name))) {
     //we copy the original file
     storageApi.PutCopy(name, tmpname, null, null, null, null,(responseMessage) => {
       if(responseMessage.code===200){
         //we append to the copy the modified doc
-        wordsApi.PostAppendDocument(tmpname, null, null, null, documentEntryList, (responseMessage) => {
-          if(responseMessage.code===200){
-            //we download the append doc
-            storageApi.GetDownload(tmpname, null, null, (responseMessage) => {
-              //we send the doc then delete it from cloud + public data
-              fs.writeFile(data_path + 'append' + name,responseMessage.body,()=>res.download(data_path + 'append' + name,null,(err)=>{
-                storageApi.DeleteFile(tmpname, null, null, (responseMessage) => {
-                  if(responseMessage.code===200) console.log('deleted');
-                  else console.log('error not deleted file '+tmpname);
-                })
-                if(err) console.log(err);
-                else fs.unlink(data_path + 'append' + name,(err)=>{
-                  if(err) console.log('err delete file in public_data');
-                });
-              }));
-            })
-          }else res.status(responseMessage.code).send(responseMessage);
-        });
+        if(wordsRegex.test(path.extname(name)))
+        {
+          const documentEntryList =  {
+            'DocumentEntries' : [{
+              'Href' : `${req.params.name}`,
+              'ImportFormatMode' : 'KeepSourceFormatting'
+            }
+          ]};
+          wordsApi.PostAppendDocument(tmpname, null, null, null, documentEntryList, (responseMessage) => {
+            if(responseMessage.code===200){
+              //we download the append doc
+              storageApi.GetDownload(tmpname, null, null, (responseMessage) => {
+                //we send the doc then delete it from cloud + public data
+                fs.writeFile(data_path + 'append' + name,responseMessage.body,()=>res.download(data_path + 'append' + name,null,(err)=>{
+                  storageApi.DeleteFile(tmpname, null, null, (responseMessage) => {
+                    if(responseMessage.code===200) console.log('deleted');
+                    else console.log('error not deleted file '+tmpname);
+                  })
+                  if(err) console.log(err);
+                  else fs.unlink(data_path + 'append' + name,(err)=>{
+                    if(err) console.log('err delete file in public_data');
+                  });
+                }));
+              })
+            }else res.status(responseMessage.code).send(responseMessage);
+          });
+        } else {
+          const appendDocumentBody = {
+          		'Document' : `${req.params.name}`,
+          		'StartPage' : null,
+          		'EndPage' : null
+          };
+          pdfApi.PostAppendDocument(tmpname, null, null, null, null, null, appendDocumentBody, (responseMessage) => {
+            if(responseMessage.code===200){
+              //we download the append doc
+              storageApi.GetDownload(tmpname, null, null, (responseMessage) => {
+                //we send the doc then delete it from cloud + public data
+                fs.writeFile(data_path + 'append' + name,responseMessage.body,()=>res.download(data_path + 'append' + name,null,(err)=>{
+                  storageApi.DeleteFile(tmpname, null, null, (responseMessage) => {
+                    if(responseMessage.code===200) console.log('deleted');
+                    else console.log('error not deleted file '+tmpname);
+                  })
+                  if(err) console.log(err);
+                  else fs.unlink(data_path + 'append' + name,(err)=>{
+                    if(err) console.log('err delete file in public_data');
+                  });
+                }));
+              })
+            }else res.status(responseMessage.code).send(responseMessage);
+          });
+        }
       } else res.status(responseMessage.code).send(responseMessage);
     })
   } else res.status(404).send("not matching extension: only pdf or word doc");
@@ -133,6 +167,7 @@ router.post('/replace-sentences',(req,res) => {
       'OldValue' : req.body.sentences.origin_sentence,
       'NewValue' : req.body.sentences.replace_sentence
     };
+    console.log(replaceTextRequestBody);
     if ( wordsRegex.test(path.extname(name))) {
       wordsApi.PostReplaceText(name, null, null, null, replaceTextRequestBody, (responseMessage) =>  {
         if(responseMessage.code===200) res.status(responseMessage.code).send("Document has been updated successfully");
